% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_funcs.R
\name{f.Weibull}
\alias{f.Weibull}
\title{Conditional density function for Weibull random variable \eqn{X} with
scalar multiplying the hazard \eqn{k}.}
\usage{
f.Weibull(x, k, parameters)
}
\arguments{
\item{x}{A vector}

\item{k}{The scalar that influences the scale
parameter for the Weibull. Consider a Weibull
with rate parameter \eqn{p} and \eqn{\lambda}.
If the distribution is re-parameterized such that
the rate parameter is \eqn{\lambda \cdot k^{p}},
then \eqn{k} proportionally scales the hazard function
for the distribution.}

\item{parameters}{A list of the two parameters
of the Weibull distribution p and lambda
before the re-parameterization described for k.}
}
\value{
The density function evaluated at x
}
\description{
\eqn{X ~ Weibull(shape=p, rate=\lambda \cdot k^{p})}
}
