% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_funcs.R
\name{f.Y.Weibull}
\alias{f.Y.Weibull}
\title{Marginal density function for conditional Weibull random variable \eqn{Y|K}
marginalized over \eqn{K}.}
\usage{
f.Y.Weibull(y, parameters, gamma_shape, gamma_rate, correlated)
}
\arguments{
\item{y}{A vector at which to evaluate the density function}

\item{parameters}{A list of the two parameters
of the Weibull distribution p and lambda
before the re-parameterization described for \code{k} in \code{f.Weibull}.}

\item{gamma_shape}{The shape parameter for gamma distribution
where in \code{f.Weibull}, \eqn{k ~ Gamma(gamma_shape, gamma_rate)}.}

\item{gamma_rate}{The shape parameter for gamma distribution
where in \code{f.Weibull}, \eqn{k ~ Gamma(gamma_shape, gamma_rate)}.}

\item{correlated}{If \code{FALSE}, then this is just a regular Weibull
density that you could get with \code{dweibull}. If \code{TRUE}, then include
the frailty parameterization marginalizing over \code{k} in \code{f.Weibull}.}
}
\description{
If \code{!correlated}: \eqn{Y \sim Weibull(shape=p, rate=\lambda)} \cr
If \code{correlated}: \eqn{Y \sim Weibull(shape=p, rate=\lambda \cdot k^{p}) \quad K \sim Gamma(a, b)}
}
