% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_funcs.R
\name{f.YZK.Weibull}
\alias{f.YZK.Weibull}
\title{Joint density function for two conditionally Weibull
random variables with shared frailty \eqn{k}.}
\usage{
f.YZK.Weibull(k, y, z, parameters_x, parameters_y, gamma_shape, gamma_rate)
}
\arguments{
\item{k}{A number or vector, the frailty term, \eqn{K = k}}

\item{y}{A number, value of \eqn{Y = y}}

\item{z}{A number, value of \eqn{Z = z}}

\item{parameters_x}{A list, with an element named \code{lambda} (\eqn{\lambda_x}), and an element named \code{p} (\eqn{p_x}).}

\item{parameters_y}{A list, with an element \code{lambda} (\eqn{\lambda_y}), and an element named \code{p} (\eqn{p_y}).}

\item{gamma_shape}{A number, the shape parameter of Gamma frailty distribution}

\item{gamma_rate}{A number, the rate parameter of Gamma frailty distribution}
}
\value{
joint density function evaluated at \code{K=k, Y=y, and X=z-y}.
}
\description{
\eqn{X \sim Weibull(shape=p_{x}, rate=\lambda_{x} \cdot k^{p_{x}})} \cr
\eqn{Y \sim Weibull(shape=p_{y}, rate=\lambda_{y} \cdot k^{p_{y}})} \cr
\eqn{Z := X + Y \implies X = Z - Y} \cr
\eqn{K \sim Gamma(shape=gamma_shape, rate=gamma_rate)}
}
