% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_funcs.R
\name{objective.function.Exponential}
\alias{objective.function.Exponential}
\title{Objective function for the Exponential deconvolution}
\usage{
objective.function.Exponential(params, y, t_y, z, t_z, control_list,
  alpha = 1, epsilon = epsilon, correlated = TRUE, int.upper = Inf)
}
\arguments{
\item{params}{A list of parameters in the order \code{lambda, mu, beta}}

\item{y}{A vector of observations representing survival at \code{t_y}}

\item{t_y}{A vector of times at which survival is observed for \code{y}}

\item{z}{A vector of observations representing survival at \code{t_z}}

\item{t_z}{A vector of times at which survival is observed for \code{z}}

\item{control_list}{A named list of options to be passed to \code{distrExIntegrate},
called \code{subdivisions} and \code{rel.tol}.}

\item{alpha}{A number, usually fixed to 1, that is the shape parameter of the frailty distribution}

\item{epsilon}{The smallest value from which to integrate, passed to integration functions, for approximations}

\item{correlated}{If \code{TRUE}, then \code{y} and \code{x = z - y} are correlated through a shared
frailty term with rate parameter \code{beta}. If \code{FALSE}, then they are independent and the function
ignores \code{beta}.}

\item{int.upper}{The largest value from which to integrate (in place of \code{Inf}) for approximations}
}
\value{
A float, the objective function evaluated at the current parameters
}
\description{
Returns the objective function evaluated with the current parameters and observed data.
Used to perform non-linear least squares. This objective function
is specific to Exponential deconvolution, so it requires 3 parameters
(one for each Weibull and one for the frailty).
}
