% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_funcs.R
\name{S.Y.Weibull}
\alias{S.Y.Weibull}
\title{Survival function for conditionally Weibull random variable \eqn{Y|K}
marginalized over \eqn{K}.}
\usage{
S.Y.Weibull(t, parameters, a, b, epsilon, correlated)
}
\arguments{
\item{t}{A number or vector at which to evaluate the density function}

\item{parameters}{A list of the two parameters
of the Weibull distribution p and lambda
before the re-parameterization described for \code{k} in \code{f.Weibull}.}

\item{a}{The shape parameter for gamma distribution
where in \code{f.Weibull}, \eqn{k ~ Gamma(gamma_shape, gamma_rate)}.}

\item{b}{The shape parameter for gamma distribution
where in \code{f.Weibull}, \eqn{k ~ Gamma(gamma_shape, gamma_rate)}.}

\item{epsilon}{A float, the lower bound for support of Y}

\item{correlated}{If \code{FALSE}, then this is just a regular Weibull
density that you could get with integrating \code{dweibull} or exact survival function.
If \code{TRUE}, then include the frailty parameterization
marginalizing over \code{k} in \code{f.Weibull}.}
}
\description{
If \code{!correlated}: \eqn{Y \sim Weibull(shape=p, rate=\lambda)} \cr
If \code{correlated}: \eqn{Y \sim Weibull(shape=p, rate=\lambda \cdot k^{p}) \quad K \sim Gamma(a, b)}
}
