% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_funcs.R
\name{S.Z.Exponential}
\alias{S.Z.Exponential}
\title{Marginal survival function for the sum
of two exponential random variables, \eqn{Z=X+Y}, with shared frailty parameter \eqn{K}.}
\usage{
S.Z.Exponential(t, a, b, parameters_x, parameters_y, correlated)
}
\arguments{
\item{t}{A number or vector, the survival times}

\item{a}{A number, the shape parameter of Gamma frailty distribution}

\item{b}{A number, the rate parameter of Gamma frailty distribution}

\item{parameters_x}{A list, with an element named \code{lambda} (\eqn{\lambda_x}), the rate
parameter of the exponential distribution}

\item{parameters_y}{A list, with an element \code{lambda} (\eqn{\lambda_y}), the rate
parameter of the exponential distribution}

\item{correlated}{If TRUE, random variables that make up the sum
are each conditionally exponential on the frailty parameter,
that has Gamma(shape=a, rate=b) distribution.}
}
\value{
Survival function evaluated at t
}
\description{
If \code{!correlated}: \eqn{X \sim Exponential(\lambda_{x}) \quad Y \sim Exponential(\lambda_{y})} \cr
If \code{correlated}: \eqn{X \sim Exponential(\lambda_{x} \cdot k) \quad Y \sim Exponential(\lambda_{y} \cdot k) \quad K \sim Gamma(a, b)}
}
