% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_funcs.R
\name{f.Weibull}
\alias{f.Weibull}
\title{Marginal density function for a Weibull
with scale parameter a function of the scalar k.}
\usage{
f.Weibull(x, k, parameters)
}
\arguments{
\item{x}{A vector}

\item{k}{The scalar that influences the scale
parameter for the Weibull. Consider a Weibull
with rate parameter \eqn{p} and \eqn{\lambda}.
If the distribution is re-parameterized such that
the rate parameter is \eqn{\lambda \cdot k^{p}},
then \eqn{k} proportionally scales the hazard function
for the distribution.}

\item{parameters}{A list of the two parameters
of the Weibull distribution p and lambda
before the re-parameterization described for k.}
}
\value{
The density function evaluated at x
}
\description{
Marginal density function for a Weibull
with scale parameter a function of the scalar k.
}
\examples{
f.Weibull(x=c(0, 0.5, 1, 5), k=0.5, parameters=list(p=1, lambda=0.5))
f.Weibull(x=c(0, 0.5, 1, 5), k=1, parameters=list(p=1, lambda=0.5))
f.Weibull(x=c(0, 0.5, 1, 5), k=4, parameters=list(p=1, lambda=0.5))
}
