% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_funcs.R
\name{S.Y.Exponential}
\alias{S.Y.Exponential}
\title{Marginal survival function for exponential
(possibly conditionally exponential)
(possibly conditionally exponential)
random variable.}
\usage{
S.Y.Exponential(t, a, b, parameters, correlated)
}
\arguments{
\item{t}{A number or vector, the survival times}

\item{a}{A number, the shape parameter of Gamma frailty distribution}

\item{b}{A number, the rate parameter of Gamma frailty distribution}

\item{parameters}{A list, with an element named lambda, the rate
parameter of the exponential distribution}

\item{correlated}{If TRUE, random variable is conditionally exponential
on the frailty parameter, that has Gamma(shape=a, rate=b) distribution.}
}
\value{
Survival function evaluated at t
}
\description{
Marginal survival function for exponential
(possibly conditionally exponential)
(possibly conditionally exponential)
random variable.
}
\examples{
S.Y.Exponential(t=0, a=1, b=1, parameters=list(lambda=1), correlated=FALSE)
S.Y.Exponential(t=c(1, 2, 3), a=1, b=1, parameters=list(lambda=1), correlated=FALSE)
}
