% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_funcs.R
\name{S.Z.Exponential}
\alias{S.Z.Exponential}
\title{Marginal survival function for the sum
of two exponential, potentially correlated random variables.}
\usage{
S.Z.Exponential(t, a, b, parameters_x, parameters_y, correlated)
}
\arguments{
\item{t}{A number or vector, the survival times}

\item{a}{A number, the shape parameter of Gamma frailty distribution}

\item{b}{A number, the rate parameter of Gamma frailty distribution}

\item{parameters_x}{A list, with an element named lambda, the rate
parameter of the exponential distribution}

\item{parameters_y}{A list, with an element named lambda, the rate
parameter of the exponential distribution}

\item{correlated}{If TRUE, random variables that make up the sum
are each conditionally exponential on the frailty parameter,
that has Gamma(shape=a, rate=b) distribution.}
}
\value{
Survival function evaluated at t
}
\description{
\eqn{X} and \eqn{Y}.
If \code{correlated}, then the distributions are
correlated through a shared frailty term, \code{k},
where \eqn{k \sim Gamma(a, b)}.
}
\examples{
S.Z.Exponential(t=2, a=1, b=1, parameters_x=list(lambda=1),
parameters_y=list(lambda=2), correlated=TRUE)
}
